+++ src/components/TemporalFilter.tsx
@@ -12,15 +12,24 @@ interface ITemporalFilterProps {
   onFromDateChange: (date: Moment) => void;
   toDate: Moment;
   onToDateChange: (date: Moment) => void;
+  timeErrorLowerBound: string;
+  timeErrorUpperBound: string;
   onClick: any;
 }
 
 export class TemporalFilter extends React.Component<ITemporalFilterProps, {}> {
   public shouldComponentUpdate(nextProps: ITemporalFilterProps) {
-    return hasChanged(this.props, nextProps, ["fromDate", "toDate"]);
+    return hasChanged(this.props, nextProps, ["fromDate", "toDate",
+      "timeErrorLowerBound", "timeErrorUpperBound"]);
   }
 
   public render() {
+    const timeError = this.props.timeErrorLowerBound || this.props.timeErrorUpperBound;
+    let timeErrorDiv = null;
+    if (timeError) {
+      timeErrorDiv = <label className="timeError">{timeError}</label>;
+    }
+
     // Notes for future CSS warriors: As of this writing, React responds to an
     // attempt to add a container around the labels and DatePicker (i.e., by
     // enclosing them in a div or section) by adding additional markup which
@@ -32,14 +41,14 @@ export class TemporalFilter extends React.Component<ITemporalFilterProps, {}> {
         <label className="from">From</label>
         <DatePicker
           id="from"
-          maxDate={this.props.toDate}
+          className={(this.props.timeErrorLowerBound === "") ? "" : "error"}
           selected={this.props.fromDate.utc()}
           dateFormat={["MM/DD/YYYY", "M/D/YYYY"]}
           onChange={(d: Moment) => this.props.onFromDateChange(d.utc())} />
         <label className="to">To</label>
         <DatePicker
           id="to"
-          minDate={this.props.fromDate}
+          className={(this.props.timeErrorUpperBound === "") ? "" : "error"}
           selected={this.props.toDate.utc()}
           dateFormat={["MM/DD/YYYY", "M/D/YYYY"]}
           onChange={(d: Moment) => this.props.onToDateChange(d.utc())} />
@@ -48,6 +57,7 @@ export class TemporalFilter extends React.Component<ITemporalFilterProps, {}> {
             <FontAwesomeIcon icon={faUndo} size="1x" />
           </button>
         </div>
+        <div className="timeError">{timeErrorDiv}</div>
       </div>
     );
   }